Split shared ovirt code into a library file
